+++ lib/rbnotes.rb
@@ -1,6 +1,8 @@
 require "textrepo"
 
 module Rbnotes
+  NAME = File.basename($PROGRAM_NAME) # :nodoc:
+
   require_relative "rbnotes/version"
   require_relative "rbnotes/error"
   require_relative "rbnotes/conf"